Update Contract Preprocessing
