+++ data/README.md
@@ -1,6 +1,10 @@
 # Data for 'Stability of LLMs for Legal Interpretation'
+## Raw Data
+`raw/vague_contracts.csv` is a copy of the data file from background work [repository](https://github.com/madiganbrodsky/vague_contracts).
+`raw/random_sample.csv` has 5 randomly selected entries with random seed of 42
 
-`vague_contracts.csv` is a copy of the data file from background work [repository](https://github.com/madiganbrodsky/vague_contracts).
-
-`random_sample.csv` has 5 randomly selected entries with random seed of 42  
+## Clean Data
+`clean/vague_contracts.csv` is the cleaned data which did the following operations
+1. Removed the bold tags (`<b>,<\b>`) meant for other purposes from the headers
+2.
